pdf2ad/requirements.txt
avro-python3==1.9.2.1
cachetools==3.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
crcmod==1.7
dill==0.3.1.1
docopt==0.6.2
fastavro==0.23.6
fasteners==0.15
Flask==1.1.2
future==0.18.2
google-api-core==1.22.2
google-apitools==0.5.31
google-auth==1.21.1
google-cloud-bigquery==1.24.0
google-cloud-bigtable==1.0.0
google-cloud-core==1.4.1
google-cloud-datastore==1.7.4
google-cloud-dlp==0.13.0
google-cloud-language==1.3.0
google-cloud-pubsub==1.0.2
google-cloud-spanner==1.13.0
google-cloud-storage==1.31.0
google-cloud-texttospeech==1.0.1
google-cloud-videointelligence==1.13.0
google-cloud-vision==0.42.0
google-crc32c==1.0.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.31.0
grpcio-gcp==0.2.2
hdfs==2.5.8
httplib2==0.17.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
libcst==0.3.10
MarkupSafe==1.1.1
mock==2.0.0
monotonic==1.5
mypy-extensions==0.4.3
numpy==1.19.1
oauth2client==3.0.0
opencv-python==4.4.0.42
pbr==5.5.0
proto-plus==1.9.0
protobuf==3.13.0
pyarrow==0.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.1
pymongo==3.11.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
six==1.15.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.10
Werkzeug==1.0.1
